pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-sender"
version = "0.1.0"
description = "Send data to Azure Event Hub and Azure IoT Hub (sync, async, batch)."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
    "azure-eventhub>=5.11.5",
    "azure-iot-device>=2.13.0",
    "azure-identity>=1.15.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[project.scripts]
azure-sender = "azure_sender.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
